feat: content security check

+++ backend/wall-server/src/main/java/cn/yiming1234/NottinghamWall/service/impl/StudentServiceImpl.java
@@ -7,6 +7,7 @@ import cn.yiming1234.NottinghamWall.dto.PageQueryDTO;
 import cn.yiming1234.NottinghamWall.entity.Student;
 import cn.yiming1234.NottinghamWall.entity.Topic;
 import cn.yiming1234.NottinghamWall.exception.LoginFailedException;
+import cn.yiming1234.NottinghamWall.exception.TeapotException;
 import cn.yiming1234.NottinghamWall.mapper.CommentMapper;
 import cn.yiming1234.NottinghamWall.mapper.StudentMapper;
 import cn.yiming1234.NottinghamWall.mapper.TopicMapper;
@@ -14,6 +15,7 @@ import cn.yiming1234.NottinghamWall.properties.WeChatProperties;
 import cn.yiming1234.NottinghamWall.result.PageResult;
 import cn.yiming1234.NottinghamWall.service.StudentService;
 import cn.yiming1234.NottinghamWall.utils.AliOssUtil;
+import cn.yiming1234.NottinghamWall.utils.ContentCheckUtil;
 import cn.yiming1234.NottinghamWall.utils.HttpClientUtil;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
@@ -42,20 +44,57 @@ public class StudentServiceImpl implements StudentService {
     private final TopicMapper topicMapper;
     private final CommentMapper commentMapper;
     private final AliOssUtil aliOssUtil;
+    private final ContentCheckUtil contentCheckUtil;
+    private String accessToken;
+    private LocalDateTime accessTokenTime;
 
     @Autowired
-    public StudentServiceImpl(WeChatProperties weChatProperties, StudentMapper studentMapper, TopicMapper topicMapper, CommentMapper commentMapper, AliOssUtil aliOssUtil) {
+    public StudentServiceImpl(WeChatProperties weChatProperties, StudentMapper studentMapper, TopicMapper topicMapper, CommentMapper commentMapper, AliOssUtil aliOssUtil, ContentCheckUtil contentCheckUtil) {
         this.weChatProperties = weChatProperties;
         this.studentMapper = studentMapper;
         this.topicMapper = topicMapper;
         this.commentMapper = commentMapper;
         this.aliOssUtil = aliOssUtil;
+        this.contentCheckUtil = contentCheckUtil;
+    }
+
+    /**
+     * 获取accessToken
+     */
+    public String fetchAccessToken() {
+        Map<String, String> map = new HashMap<>();
+        map.put("appid", weChatProperties.getAppid());
+        map.put("secret", weChatProperties.getSecret());
+        map.put("grant_type", "client_credential");
+        String json = HttpClientUtil.doGet(WX_ACCESS_TOKEN_URL, map);
+        JSONObject jsonObject = JSON.parseObject(json);
+        log.info("access_token:{}", jsonObject.getString("access_token"));
+        // 保存accessTokenTime （有效期限）（now()+7200）
+        accessTokenTime = LocalDateTime.now().plusSeconds(jsonObject.getLong("expires_in"));
+
+        accessToken = jsonObject.getString("access_token");
+        return accessToken;
+    }
+
+    public String getAccessToken() {
+        if (accessToken == null) {
+            log.info("token为空，重新获取");
+            return fetchAccessToken();
+        } else if (accessTokenTime.isBefore(LocalDateTime.now())) {
+            log.info("token失效");
+            return fetchAccessToken();
+        } else {
+            log.info("token未失效,直接返回");
+            return accessToken;
+        }
     }
 
     /**
      * 获取openid
+     * @param code 微信登录凭证
+     * @return openid
      */
-    private String getOpenid(String code) {
+    public String getOpenid(String code) {
         Map<String, String> map = new HashMap<>();
         map.put("appid", weChatProperties.getAppid());
         map.put("secret", weChatProperties.getSecret());
@@ -69,21 +108,9 @@ public class StudentServiceImpl implements StudentService {
     }
 
     /**
-     * 获取accessToken
-     */
-    private String getAccessToken() {
-        Map<String, String> map = new HashMap<>();
-        map.put("appid", weChatProperties.getAppid());
-        map.put("secret", weChatProperties.getSecret());
-        map.put("grant_type", "client_credential");
-        String json = HttpClientUtil.doGet(WX_ACCESS_TOKEN_URL, map);
-        JSONObject jsonObject = JSON.parseObject(json);
-        log.info("access_token:{}", jsonObject.getString("access_token"));
-        return jsonObject.getString("access_token");
-    }
-
-    /**
      * 微信登录
+     * @param studentLoginDTO 学生登录DTO
+     * @return 学生
      */
     @Override
     public Student wxLogin(StudentLoginDTO studentLoginDTO) {
@@ -123,6 +150,9 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 微信获取手机号
+     * @param code 微信登录凭证
+     * @param id 学生id
+     * @return 手机号
      */
     @Override
     public String getPhoneNumber(String code, Integer id) throws IOException {
@@ -148,6 +178,8 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 更新手机号
+     * @param id 学生id
+     * @param phoneNumber 手机号
      */
     @Override
     public void updatePhoneNumber(Integer id, String phoneNumber) {
@@ -161,11 +193,25 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 更新学生信息
+     * @param studentDTO 学生DTO
+     * @return 学生
      */
     @Override
     public Student update(StudentDTO studentDTO) {
         Student student = studentMapper.getById(studentDTO.getId());
         if (student != null) {
+            String accessToken = getAccessToken();
+            boolean isUsernameSafe = contentCheckUtil.checkTextContent(studentDTO.getUsername(), 1, student.getOpenid(), accessToken);
+//            boolean isAvatarSafe = contentCheckUtil.checkImageContent(studentDTO.getAvatar(), 1, student.getOpenid(), accessToken);
+            if (!isUsernameSafe) {
+                log.info("username:{}", studentDTO.getUsername());
+                throw new TeapotException(MessageConstant.CONTENT_UNSECURED);
+            }
+//            if (!isAvatarSafe) {
+//                String objectName = studentDTO.getAvatar().substring(studentDTO.getAvatar().lastIndexOf("/") + 1);
+//                aliOssUtil.delete(objectName);
+//                throw new TeapotException(MessageConstant.CONTENT_UNSECURED);
+//            }
             if (!student.getAvatar().equals(studentDTO.getAvatar()) && !student.getAvatar().contains("default.jpg")) {
                 String objectName = student.getAvatar().substring(student.getAvatar().lastIndexOf("/") + 1);
                 aliOssUtil.delete(objectName);
@@ -183,6 +229,8 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 根据id查询学生
+     * @param id 学生id
+     * @return 学生
      */
     @Override
     public Student getById(Integer id) {
@@ -191,6 +239,8 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 根据用户名查询学生
+     * @param username 用户名
+     * @return 学生
      */
     @Override
     public Student getByUsername(String username) {
@@ -199,6 +249,8 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 根据学号查询学生
+     * @param studentId 学号
+     * @return 学生
      */
     @Override
     public Student getByStudentId(Integer studentId) {
@@ -207,6 +259,8 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 根据邮箱查询学生
+     * @param email 邮箱
+     * @return 学生
      */
     @Override
     public Student getByEmail(String email) {
@@ -215,6 +269,8 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 学生分页查询
+     * @param pageQueryDTO 分页查询DTO
+     * @return 学生分页结果
      */
     public PageResult pageQuery(PageQueryDTO pageQueryDTO) {
         PageHelper.startPage(pageQueryDTO.getPage(), pageQueryDTO.getPageSize());
@@ -227,6 +283,10 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 查询发布的帖子（分页）
+     * @param id 学生id
+     * @param page 页码
+     * @param pageSize 每页大小
+     * @return 发布的帖子
      */
     @Override
     public PageResult<Topic> getPublishedPosts(Integer id, int page, int pageSize) {
@@ -239,6 +299,10 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 查询收藏的帖子（分页）
+     * @param id 学生id
+     * @param page 页码
+     * @param pageSize 每页大小
+     * @return 收藏的帖子
      */
     @Override
     public PageResult<Topic> getCollectedPosts(Integer id, int page, int pageSize) {
@@ -252,6 +316,10 @@ public class StudentServiceImpl implements StudentService {
 
     /**
      * 查询评论的帖子（分页）
+     * @param id 学生id
+     * @param page 页码
+     * @param pageSize 每页大小
+     * @return 评论的帖子
      */
     @Override
     public PageResult<Topic> getCommentedPosts(Integer id, int page, int pageSize) {
